+++ package.json
@@ -46,6 +46,11 @@
     "preact-render-to-string": "^5.1.12",
     "react": "npm:@preact/compat",
     "rehype-accessible-emojis": "^0.3.2",
+    "rehype-plaintext": "^1.0.0",
+    "rehype-raw": "^5.0.0",
+    "remark": "^13.0.0",
+    "remark-mdx": "^1.6.22",
+    "remark-mdx-to-plain-text": "^3.0.0",
     "remark-shiki-twoslash": "^1.0.1",
     "remark-slug": "^6.0.0",
     "rollup": "^2.0.0",
