+++ src/components/Blog/BlogPost.tsx
@@ -1,14 +1,18 @@
-import { seo }                from 'microsite/head';
-import { VNode }              from 'preact';
-import { Head }               from '@/components/Head';
-import { Main }               from '@/components/Main';
-import { Timestamp }          from '@/components/Timestamp';
-import { TopicTagList }       from '@/components/Topic/TopicTagList';
-import { Topic }              from '@/lib/content';
-import { styled }             from '@/lib/styles';
-import { BlogArt }            from './BlogArt';
-import { BlogArtDefs }        from './BlogArtDefs';
-import { FullBleedContainer } from '../FullBleedContainer';
+import { seo }                 from 'microsite/head';
+import { ComponentChildren }   from 'preact';
+import { FullBleedContainer }  from '@/components/FullBleedContainer';
+import { Head }                from '@/components/Head';
+import { Main }                from '@/components/Main';
+import { Timestamp }           from '@/components/Timestamp';
+import { TopicTagList }        from '@/components/Topic/TopicTagList';
+import {
+  MDXDescription,
+  Topic,
+} from '@/lib/content';
+import { styled }              from '@/lib/styles';
+import { BlogArt }             from './BlogArt';
+import { BlogArtDefs }         from './BlogArtDefs';
+import { BlogPostDescription } from './BlogPostDescription';
 
 const BlogPostTitle = styled('h1', {
   marginBottom: '0.25rem',
@@ -19,8 +23,8 @@ const BlogPostHeading = styled(FullBleedContainer, {
 });
 
 interface BlogPostProps {
-  readonly children?:   VNode;
-  readonly description: string;
+  readonly children?:   ComponentChildren;
+  readonly description: MDXDescription;
   readonly hash:        string;
   readonly stat: {
     readonly created: Date;
@@ -31,7 +35,10 @@ interface BlogPostProps {
 
 export const BlogPost = ({
   children,
-  description,
+  description: {
+    Component: Description,
+    raw:       description,
+  },
   hash,
   stat: { created },
   title,
@@ -50,10 +57,14 @@ export const BlogPost = ({
         <BlogArtDefs />
         <BlogArt hash={ hash } title={ title } topics={ topics } />
 
-        <BlogPostTitle>What the art?</BlogPostTitle>
+        <BlogPostTitle>{ title }</BlogPostTitle>
         <Timestamp date={ created } itemprop="datePublished" />
         <TopicTagList link={ false } topics={ topics } />
       </BlogPostHeading>
+
+      <BlogPostDescription>
+        <Description />
+      </BlogPostDescription>
 
       { children }
     </Main>
