+++ typings/remark-mdx-to-plain-text copy/index.d.ts
@@ -1,0 +1,7 @@
+declare module 'remark-mdx-to-plain-text' {
+  import { Plugin } from 'unist';
+
+  const remarkMDXToPlainText: Plugin;
+
+  export default remarkMDXToPlainText;
+}
