+++ typings/remark-mdx/index.d.ts
@@ -1,0 +1,7 @@
+declare module 'remark-mdx' {
+  import { Plugin } from 'unist';
+
+  const remarkMDX: Plugin;
+
+  export default remarkMDX;
+}
